matcher/pattern/file/serialized: move File into it's own module
